moved employeed data out of the router file
